requirements.txt
flask 
pymysql
python-dotenv
secure-smtplib
schedule
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client 
psycopg2-binary